(Mostly) ignore whitespace in the lexer
Ignore it everywhere except for inside PRE_FENCE blocks.

This feels like the right thing to do, as the AST mostly shouldn't care
about any of these. Having whitspace riding along as a stowaway on other
tokens felt like a smell.

I am super tired and the code is a bit messy in places, but you can see
plenty of bits and pieces where this is clearly better (ie. fewer
`trim()` calls and such). I'll look to clean this up further as I go.