Add debug switch
Without it, we show abbreviated error messages. With it, we'll let the
full stack trace show (which doens't happen to be very intersting for
the ENOENT errors that I tested this with -- ie. `node dist/index.js -C
some-non-existent-dir` -- so I might have to look into how/why detail
might be getting swallowed there).