Add starting point for initial integration test
This is the thing which triggered the move from Jest to Mocha. Jest was
segfaulting as soon as I trying to import the bluebird module, despite
the presence of the manual mock.

Even if I could work out the right voodoo incantation to make that work,
I prefer to just jettison the magical `require` machinery entirely and
have something predictable.

Actual tests to follow.

Note: the filename here probably sucks. Will change it later if
something better occurs to me.