Tear down Sinon stubs automatically after each example
Maybe this is a horrible way to do this, but I suspect it beats the
annoyance of forgetting to clean up a stubbed method.

We create a sandbox at the beginning of each example and tear it down
afterwards (http://sinonjs.org/docs/#sinon-sandbox), doing so using the
"root-level hooks" mentioned in the Mocha docs (http://mochajs.org/). We
expose the sandbox to the tests via a `stub` global, which brings the
number of globals to about three:

- sinon
- expect
- stub

With the evident possibility of adding a couple more (`mock`, `spy`)
later.

Another alternative here is I could export my own "sinon" global which
ends up proxying to the real sinon under he covers. I'll think about
that separately.