Add Mocha "WatchReporter"
This is just like the "min" reporter that comes with Mocha, but it
doesn't clear the screen.

We don't want to clear the screen because the test run is happening in
parallel with other checks (linting, building, type-checking), so we
don't want any output from those other jobs getting blown away before
you have a chance to read it.