Parse function "attributes"
`abort`, `range`, and `dict`.

Like Vim, we allow repetition, although it is kind of non-sensical, and
we also require the `:function` command to be on a line by itself.

Note the use of `sepEndBy` rather than `sepBy`, otherwise any trailing
whitespace will be taken to indicate that another attribute follows,
even though only a newline may follow.