Require end-of-statement after :endf
Allows us to parse a trailing "endf\n".

Note that `:h :endf` says `:endf` must be on a line all by itself, but
this is not technically true. Vim will happily accept this:

  endf | echomsg "Hi!"

Even if it doesn't actually run the `:echomsg` call... It does, however,
complain with:

  E126: Missing :endfunction

if you try to give it something like:

  echo 1 | endf

I'm currently not handling that case, I think.