Strip diff headers
This is a gross hack, probably at the wrong layer of abstraction (ie.
inside `goldenVsStringDiff'`, which supposedly doesn't know about the
program being used to make the diff).

Note that I had to temporarily turn off color in order to get my "@@"
prefix test to work (because otherwise we have a color escape sequence
at the start of the line with prevents it from being considered a
suffix).

Will think about whether I want to keep this in.