Trim trailing whitespace on headings
Horribly inefficient but easy-to-read `strip` implementation taken from:

http://stackoverflow.com/questions/6270324/in-haskell-how-do-you-trim-whitespace-from-the-beginning-and-end-of-a-string

The commented out bits are a sketch of a lookahead-based alternative
that I was in the process of writing when it occurred to me that I could
strip instead. I think I will go ahead and do that anyway, but for now I
wanted to put this in here to reflect my intermediate state.