Make blockquote handle phrasing content too
Note that we are allowing empty blockquotes:

  "" >

Which is a bit dodgy, and also not parsing paragraphs:

  ""
  " > a
  " >
  " > b

is treated as one big paragraph ("a", Whitespace, "b"). Ideally, I'll
fix both of those issues. Depending on how ugly that gets, I can
consider allowing other things to be nested inside blockquotes, like
lists and code fenced sections.