Add "integration" test
This is a snapshot of the current vim-docvim documentation. Note that I
had to add another dastardly `try` to get this to parse, thanks to my
`<greg@hurrell.net>` email address, which was tripping us up trying to
parse br tags.

The AST looks pretty reasonable at this point, and we can evidently
handle well-formed input quite well.