Add a Monoidal tree walker
I built this in a couple of steps, counting nodes:

- With concrete Int type.
- With general "a" type.
- With a "Monoid a" constraint, given a Sum instance.

Finally, the version committed here collects numbers instead of summing
them, showing that this can be used with the list monoid.

I am not sure what the most useful thing is going to be here. If I
assume Monoid, then we can probably implement this in terms of `mappend`
and `mconcat` etc with maybe even a default walker instance.

But on the other hand, I know I am going to want to do things like build
up symbol tables, and I am not sure if I want to do that with a list
(probably a list of "Mabye LinkTarget" or something, which I can then
filter and sort for dupes and then turn into a set).