Use symbol table to conditionalize Link printing behavior
Stuff we have in our symbol table becomes an internal link. Stuff we
haven't seen is probably a Vim help target outside our control (or a
error on the part of the documentation author, but we can't know that),
and we render it, for now, using backticks.

This introduces a couple of test failures, but I am not fixing them yet;
need to decide what I want the actual behavior here to be. Perhaps
backticks will be fine.