Handle more kinds of let assignment
Huge kludge territory at this point, until we have a proper expression
parser, but working through the Ferret fixtures I discovered the need to
parse:

  let l:list[l:line - 1]=0

Note that in adding the test cases I discovered a problem with our
trailing comment parsing. This would parse fine:

  let g:foo=1 " Trailing.
  unlet g:foo

But this would not:

  let g:foo=1 " Trailing.

  unlet g:foo

Because our comment parse needed to slurp up the empty lines.