Add support for lexpr, lw
Notably, this is enough to make the entire Ferret fixture parse without
blowing up. There are definitely bugs in here (just look at the ASTs
produced) that will need to be fixed, but it's nice that it doesn't
explode.

I am probably doing this l-statement disambiguation in a stupid way
(probably need to start thinking about leveraging the token parser
stuff), but this works so a I am going to run with it.