Fix overly eager GenericStatement parsing and eos issues
Too much stuff was getting parsed as generic statement. Here we fix that
and the related issue of "lw 100 f" being parsed as "lw 100" + "f"
(GenericStatement).

This does introduce a slight regression in that trailing comments are
now considered part of the RHS of a let statement, but it still fixes
more than it breaks, so I am keeping this in.

If I get to proper expression parsing (or at least proper string
literals) then I'll be able to fix that.