Traverse into functions in Markdown printer
Wondering if I should be able to re-use `walk` here. Not sure how easy
it will be to get the Reader monad and the `walk` function to play
nicely together. I think it will be worth figuring it out though, as
I'll need to make a bunch of pruned AST trees in order to produce the
different document sections (mappings, options etc) so otherwise will
wind up writing a lot of similar code in walkers for each one.