(Temporarily) switch to partitioning footer visitor
This is semi-broken, so will revert, but I wanted to show where my
thinking erroneously took me.

The main idea is that if I extract the footer sections from the tree,
then I am left with "something" that I need to pass to the Markdown
printer and that "something" should _not_ have the extracted contents in
it as well, otherwise they will appear as duplicated in the output.

So, I thought to partition instead of extract. One list would be
extracted footer nodes and another whatever was left. I could then
string these extractors into a pipeline, effectively filtering the
remaining output down as I went. Or I could merge them into one big
extractor/classifier which would return a collection of non-overlapping
subsets.

This is flawed for two reasons:

(1) My implementation was wrong. As committed here, the partitioning is
half-busted because of my `nodes` implementation. I tried a few
different ways, only some of which type-checked and none of which
actually produced the right results (either too much was accumulated,
producing repeated entries, or not enough).

(2) The conceptual model is fundamentally flawed because what I need is
a tree pruning process and not something that can only output arrays
(even if those arrays contain trees). This was my principal mistake,
because I thought this could be made to work because we're dealing with
subtrees. Maybe it can, but I don't think so: a `Unit` might be a list
`Node`s, but all the interesting distinctions happen inside some nested
child (usually a DocBlock, but not always: for example it might be
inside a DocBlock inside a FunctionDeclaration).

So, I am going to back this out and reconsider my approach.

Footer extraction will still work but I need to figure out what to do
with what's left, and whether I can really do this properly by pruning
at a tree level.