Update integration test files
Evidently, we have a circular dependency here. You update the
input files, get new output, copy the output over to the
other repo, commit it there, and now your input files have
changed. Not too worried about this, but it would be nice to
find a neat way to avoid the issue.