Don't use -Werror in Stack
I intend this to mostly be a local feedback tool, so I *think* that defaulting
this to off is right call. I fixed the specific build failure cited below anyway
in c9efe9d7b010adbd1, but I think it's ok to turn off `-Werror` in Stack because
other "real" errors will still be errors; these are just warnings.

I've added explicit `-Werror` to all my wrapper scripts except `bin/repl`, which
is not "build-like" and I want to be lenient. (I actually meant to make it more
lenient in the future by deferring type errors with `-fdefer-type-errors`.

See:

- https://github.com/wincent/docvim/issues/24
- https://travis-ci.org/wincent/docvim/jobs/138094062
- https://www.reddit.com/r/haskell/comments/4np0l2/a_ghci_annoyance_wondering_if_anyones_else_also/