fix: work around GitHub formatting issue
When we generate Markdown like this:

    <p align="right"><a name="loupe-smartcase-override" href="#user-content-loupe-smartcase-override"><code>loupe-smartcase-override</code></a></p>
    <strong>`'smartcase'`</strong>

    Turned on (overrides <strong>`'ignorecase'`</strong>, making the search pattern case-sensitive whenever it containers uppercase characters).

GitHub mangles the first `<strong>` element:

    <strong>`'smartcase'`</strong>

But does the right thing with the second one:

    <strong><code>'ignorecase'</code></strong>

If we add another line break in between, the problem goes away.

So, rejig how we emit newlines. Instead of emitting them before
headings, emit them after link targets. This ends up looking better
anyway, resulting in fewer blank lines overall in the output.

Sample change in Loupe project as a result of this change:

https://gist.github.com/wincent/3305bd10ba651320b2c3ea4852d22e91