refactor: split into main.rs + lib.rs
I don't want to go crazy with over-abstracting this at the early stages,
but I do want to be able to run unit tests easily, so it makes sense to
put most of the code in a library crate. "main" won't really do anything
other than pass in options to the library and wait for it to do its
magic.