refactor: start playing around with TokenBuilder
This is increeeedibly ugly, but I need to figure out how to pack token
contents into some token types. Lots of copying and mutable borrowing
that I am not sure I am doing in the best possible way. All of this
because I realized I can't cheaply slice back into the input to get the
token contents, due to the lack of cheap index-based access to strings.