refactor: switch everything over to "cheap" model
Lots of hoop-jumping required though. As explained in depth here:

    https://lukaskalbertodt.github.io/2018/08/03/solving-the-generalized-streaming-iterator-problem-without-gats.html

you can't have lifetimes in an implementation of the Iterator trait (if
you add them, you're no longer implementing the trait). Similar points
are made here:

    https://stackoverflow.com/questions/30422177/how-do-i-write-an-iterator-that-returns-references-to-itself

So, I was only implementing the trait so I could use `collect()` in the
tests. Figured I'd just write my own `collect()` then. But it ended up
getting _super_ messy, as you can see. The naive implementing ends up
running into mutable borrow problems; ie. if you call
`lexer.next_token()` in a loop, Rust complains about repeated mutable
borrows.

The end result is what you see here: carefully "cloning" the returned
values by hand so that Rust doesn't catch you returning anything that
came from the lexer itself. I'm ok with this only because it's in the
test suite.

But this also hints at problems we're going to have in the parser: it's
currently working only because it just calls methods on the lexer; it
doesn't tend to return any tokens (or anything from tokens) to higher
levels. This suggests to me that when we return an AST; the AST will
need a copy of any string fragments that we return -- I don't think
we're going to be able to use references to the original slice. All of
which is a bit of a shame because it would be really nice if we could
leverage the immutability of the input string and just slice into it
without any copying for the entire program.