refactor: make Peekable non-generic
This is way tidier, albeit less flexible (do we really need that
flexibility though?). Instead of wrapping the std Peekable, we assume
we're working with chars. So we are changing our `new()` function to
take a str instead of an iterator, we construct our own iterator, we
manage byte and character indices, and we return `Option<char>` instead
of `Option<&SomeGenericThing>` (ie. `Option<&char>` in our case).