refactor: switch back to using a real Iterator
I need to be able to peek ahead at lexer tokens (ie. with peekable) in
order to do anything useful in the parser. Well, not strictly "need",
but it will make my life easier.

So, instead of storing token text with some tokens, store byte indices
so I can extract the slices out at will (tests show that I am probably
going to be able to do that just fine).