refactor: flatten AST node types
Switch one-field structs like:

    Chunk { block: Block }
    Block { statements: Vec<Statement> }

to tuple stucts (ie. structs with anonymous fields):

    Chunk(Block)
    Block(Vec<Statement>)

This has the obvious upside of making the AST representation more
compact (ie. _displays_ well), with the downside that tuple-based
_access_ doesn't read so well; eg.

    Before: self.ast.block.statements.push(...)
    After:  self.ast.0.0.push(...)

Going to run with it and see how it pans out.

The other issue is that with `Name` and `Number` becoming tuple structs,
we can no longer distinguish them "for free" from the token structs of
the same name. So, I had to switch to referring to the lexer token
structs with qualified names. I think in the long term I have to figure
out a way to tag all Lexer types in the parser; maybe some type aliases
would suffice.