refactor: hoop jump to make Peekable wrapper more "perfect"
"Perfect" in the sense of perfectionism; I didn't like that the real
peekable returns an Option of a reference to its Item type, and I was
returning the non-reference value directly.

Note that this isn't quite as efficient as the real peekable (or maybe
it is), because the real peekable uses `get_or_insert` (ie. it only
replaces a `None`, but not a `Some`, whereas we just replace the `Some`
no matter what). I don't know how many repeated calls to `peek()` we can
really expect though.