feat: improve string handling
But not really, because as you can see I got defeated by the
borrow-checker at the end.

Basically, I want to make a new str slice that has all the string
escapes removed. So I make an owned string and remove them. But then I
can't return the string (the AST type expects a slice, not a string),
and I can't return a reference either, because the string will get
dropped on exiting the function.

Not sure how to fix that without changing the type of the AST, which
seems undesirable, but may be the only way... Maybe these strings need
to be Boxed to show that they may be dynamically allocated.