refactor: split string nodes into two types
Cooked (ready to eat) and Raw (still has escape sequences literally
present in it).

eg. Raw: "hello\\nworld"

vs: Cooked: "hello
world"

Because the Raw string is unchanged, it can just use a string slice to
reference the input string.

The Cooked string, on the other hand, must be dynamically allocated, and
lives on the heap.