refactor: extract cook_str method
I might need to "cook" in more than one place, but that's not the reason
I'm doing this. The real motivation is that I want to remove what is
relatively extraneous detail from the parse_exp() method to keep it
readable: it is going to get more complicated as I implement operator
precedence and associativity.