refactor: move error modules into separate files
Just to declutter. In general, I don't like splitting things into files
too prematurely, especially when the details are changing fast. But the
basic structure of these error types seems relatively stable at this
point and I seldom have to look inside the, so this seems like a
reasonable change.