feat: make it almost work
Shamelessly copying from:

    http://simplygenius.net/Article/DiffTutorial1#heading0011

and then beating it with a hammer until it almost works.

I couldn't figure out from the paper how to reconstruct the darn edit
graph. A classic case of "we simply do X" and I had not idea what that
meant. Specifically, it says:

> To explicitly generate a solution path, O(D^2) space is used to store
i> a copy of V after each iteration of the outer loop. Let Vd be the
> copy of V kept after the d th iteration. To list an optimal path from
> (0,0) to the point Vd [k] first determine whether it is at the end of
> a maximal snake following a vertical edge from Vd − 1 [k + 1] or a
> horizontal edge from Vd − 1 [k − 1]. To be concrete, suppose it is
> Vd − 1 [k − 1]. Recursively list an optimal path from (0,0) to this
> point and then list the vertical edge and maximal snake to Vd [k].
> The recursion stops when d = 0 in which case the snake from (0,0) to
> (V0 [0],V0 [0]) is listed. S

Differences from that web solution:

- Using recursion like in the paper instead of iteration.
- I have a bounds overflow 🥳 — no idea why or how to fix.
- I am starting at `d` not `d-1`.
- ???