test: show why we need to refactor recursive solution to be iterative
This one may end up being nothing more than a shadow of a memory in the
Git history, because I am going to refactor the recursion away, and I
don't see any point in keeping the original one around at that point.

Of course, you'd be mad to use this implementation for inputs like this
anyway: 10k deletions followed by 10k insertions gives us an edit script
of 10000 + 10000 long, which not only means 20000 stack frames to
reconstruct the edit path (which is why we're overflowing the stack) but
10k * 10k * 20k space for all the `v` vectors... that's 2,000,000,000,000
"items", and will serve as motivation to implement the linear space
version next.