fix: fix some edge cases in find_middle_snake
This still doesn't work properly though.

The `println!` statements and the tests show me that it finds the middle
snake in the example input just fine. When we recurse on the first
subproblem, though, things go awry.

- It computes the forward D-paths for d=0.
- It computes the backward D-paths for d=0.
- It computes the forward D-paths for d=1.
- It computes the backward D-paths for d=1; so far so good.
- It computes the forward D-paths for d=2, and BOOM!\

You see this output:

    forward search d=2 (of 3) k=-2 (of 2)
    x -> 0 (down)
    forward search d=2 (of 3) k=0 (of 2)
    x -> 1 (down)
    x -> 2 (snake)

So, it knows that on k - 2 it's on the left edge and can only have
gotten there from above; x is stuck on 0 - ie. this is the vertical move
from (0, 1) to (0, 2).

And it knows that on k - 0 it's in the middle and it has a juicy
downwards path coming from above at x = 1; furthermore, it has a snake
which gets it all the way to x = 2 - ie. this is the vertical move from
(1, 0) to (1, 1), and the snake from there to (2, 2).

It then checks for overlap, finds it, and returns, before it has had a
chance to investigate the even better path that it would have found if
it had gone right from (1, 0) to (2, 0), followed by a snake down to (3,
1) (ie. landing us on x = 3).

Looking at the paper itself, it doesn't show us how to check for
overlap, but it does show the overlap check happening inside the inner
for loop. It is a greedy algorithm, after all. Likewise, the apparently
working implementations I have been poring over online all seem to do
the check inside the inner loop. So it's not clear how we would get the
algorithm to pick the better path here...

I can cheat and reverse the iteration order, forcing it to look at
k-lines in order +2, 0, -2, and that seems to make it work for one
example, at the cost of breaking another.

I think the problem might be that my reverse search need to optimize for
insertion (vertical moves), not deletion (horizontal moves), in order
to make it meet the forwards search where I want it to. I'll look at
that after this. Looking at implementations, I see one that reverses
both loops, another that reverses neither, and also reverses neither but
might implement the "upwards" bias that I'm talking about here.