fix: correct remaining bugs
I think...

Still some comments in here I would like to polish up, and also the darn
thing is slow for worst-case huge inputs (eg. n = 10,000, m = 10,000,
with 10k deletions and 10 insertions required).

The one bit I don't fully grok is the ruling out of "impossible"
overlaps in the `find_middle_snake` function. Going to think a bit about
that and clarify it if I can.