perf: avoid repeated allocations
Another experiment that I am going to commit and revert just so that it
is in the history. Wanted to see if I could make the pathological case
better by avoiding repeated allocations and recycling my ring buffers,
but it doesn't make any difference and obviously makes the code less
readable.

Will need to do something cleverer like Git does:

    https://github.com/git/git/blob/66262451ec94d30ac4b80eb3123549cf7a788afd/xdiff/xdiffi.c#L142-L151