perf: cache string hashes for 2x speed-up
On this machine, swapping the hash implementation dropped our test run
for the pathological input case (2 x 10k files, 10k deletions, 10k
insertions) from 33.8s to 27.97s (a roughly 18% speed-up).

In this commit we drop it down to 14.14s (an _additional_ 2x speed-up)
by caching hash values instead of recomputing them over and over again
as we subdivide the problem during `find_middle_snake` recursion.

Code is ugly, so consider this a proof of concept, and also note that we
are naughty for using hashes in this way anyway.

For example, this splendid Stack Overflow answer:

    https://softwareengineering.stackexchange.com/a/145633

reports that for a list of 216k English words, our hash function (djb2a)
finds 5 collisions. For a list of 215k UUIDs, it finds 6 collisions. For
the numbers 1 through ~216k it finds 0 collisions.

Concretely, and just for fun, the word collisions are:

> - `haggadot` collides with `loathsomenesses`
> - `adorablenesses` collides with `rentability`
> - `playwright` collides with `snush`
> - `playwrighting` collides with `snushing`
> - `treponematoses` collides with `waterbeds`

This is not to say that this makes it a bad hash function. Even a
well-known "great" function like Murmur has 6, 5, and 0 collisions
respectively for the same input (albeit with better apparent
randomness).

All of this is to say that I have to switch to using hashes for negation
only, and inputs which might match will then need to be subjected to a
proper equality check; hopefully that won't slow things down too much.