chore: add script for detecting hash collisions
These are the dupes for the original 32-bit djb2 hash function:

    joyful          121541278    synaphea     121541278
    dram            2090194761   vivency      2090194761
    depravement     218169280    serafins     218169280
    stylist         2945051873   subgenera    2945051873
    hetairas        3940087062   mentioner    3940087062
    redescribed     555028097    urites       555028097
    heliotropes     874715923    neurospora   874715923

and the modified 32-bit djb2a function:

    playwright      195669366  snush           195669366
    treponematoses  2371843200 waterbeds       2371843200
    adorablenesses  2661994410 rentability     2661994410
    haggadot        379918194  loathsomenesses 379918194
    playwrighting   908587542  snushing        908587542

except for the fact that both "outsourcing" and "outsourcings" actually
appear twice in the input file, so are spurious collisions and I've
omitted them.

As expected, both of the above sets match the collisions listed here:

    https://softwareengineering.stackexchange.com/a/145633

Here I would show you the collisions for the 64-bit variants, but there
are none, except for the spurious ones... ie.

    outsourcing   13888261722060976679  outsourcing   13888261722060976679
    outsourcings  15590779058982991738  outsourcings  15590779058982991738

for djb2... and:

    outsourcings  11842848437052154496  outsourcings  11842848437052154496
    outsourcing   13774688066850951315  outsourcing   13774688066850951315

for djb2a. As they say, we're gonna need a bigger boat...