refactor: simplify types
I thought the `Index` trait was pretty "clever", but the truth is I
would like to require something that implements `iter()` as well (for
ease of construction of the hashes vector. There is no trait for that.
There is `IntoIterator`, which Vec implements. Maybe I should just ask
for a darn Vec.