docs: write up what look to be some bugs in upstream histogram implementations
I'm not confident about this because the upstream projects are of
legendary quality, but I have been looking at these quite closely and I
have yet to find a mistake in my reasoning. May be worth a message to
the Git mailing list to see if anybody can confirm the analysis.