fix: fix histogram diff
Note the remaining (known) bug in histogram diff was actually a "bug" in
the Myers diff. This should be a "replace" edit. So, implementing it
locally for now, and added a failing test case to see if we can get
Myers to behave better for that case in a future change.