feat: add first pass at printing proper context and hunk headers
As noted, I almost certainly have an off-by-one (or more) in here, but
this produces reasonable output for our histogram test case diff. Going
to add more tests, obviously, and fix any bugs that I find, in future
commits. Also as noted inline, probably want to extract this into
another file.

And then once all that is done, look at refactorings. For now I've
written it "dumb and verbose" just to see it working, but there may be
patterns in here that can be used to reduce the duplication.