refactor: prepare to DRY up snapshot testing
Wanted to use a procedural macro for this but there is currently no way
with the stable compiler (have to use the nightly).

Basically, need this feature:

    https://github.com/rust-lang/rust/issues/54725

as described here:

    https://stackoverflow.com/a/65813587/2103996

I want to know where the macro is used, not where it is defined, so
that I can do:

    #[check_snapshots]
    fn transform() {...}

And have it generate a bunch of tests:

    fn transform() {...}

    #[test]
    fn test_first_snapshot() {...}

    #[test]
    fn test_second_snapshot() {...}

    // etc...

But I only have a stable compiler installed. So, will have to go back to
the old "macro_rules!" approach, which means running all the tests in a
single test function (sadface), and that in turn means the first failure
stops all the others from running.

Either that, or stick with the old pattern of listing them all out one
by one...

Another option I considered was passing in attributes like this:

    #[check_snapshots("path_here")]
    fn transform() {...}

but that would involve some ugly TokenStream parsing, and I don't want
to pull in a dependency for that (although, I guess it wouldn't be the
end of the world to write out something crude for this...).