feat: show how we could do check_snapshots macro
Just showing this for the record; I have no intention of keeping it.

Note that a procedural macro crate may contain _nothing_ else other than
the macro, so in order to actually use this I'd need to split it into a
separate package. I can't have two crates in the same package and
combine them without the compiler complaining.

In short, so ridiculously fragile and cumbersome that I think I'll just
forget about it and go back to spelling things out manually.