refactor: avoid concurrency issues with dry_run param
I looked into setting up a global mutex to allow this one test to run
alone, but it is fraught with complication by the looks of things. Let's
just suffer some additional API complexity in order to side-step the
problem. The ironic thing is that globals are hard to do in Rust because
globals are considered evil (why make it easy?), but my use of
environment variables in the first place was where the original
fall-from-grace happened.

I am still semi-interested in doing it the hard way, just as a learning
exercise.