feat: parse `return` and `break`
Although as noted in comments, nothing in here is explicitly forcing
that these be the last statement in a block. I haven't checked yet, but
it's possible this kind of falls out naturally from the rest of the
system anyway (ie. after the block, if there is a statement other than
"end" etc then we might get an `UnexpectedToken` error anyway).