refactor: move away from `Box<dyn Error>` in parser
Because I've discovered that this makes errors really easy to throw, but
not so easy to consume (eg. you don't know you've got a `ParserError`,
so you can't access `position` on it). So, use `std::convert::From`
trait so that we can wrap other errors (eg. `LexerError`) by turning
them into `ParserError`, then git rid of `Box<dyn ...>`.

Still need to decide what to do with those two ugly `Utf8Error` and
`ParseIntError` ones because we don't have positions at that point; will
probably just change the type to `Option` so that position can be `None`
(ie. unknown) for those.