test: add corpus.snap to parser tests as stress test
Now that the parser is close to being "complete" (in the sense that it
can in theory parse most valid Lua), let's see how far it can get
through the 600+ lines of Lua in the main Corpus Lua file.

Just over 50 lines in... not bad, I guess... We die because I failed to
correctly implement the very first variant of the `stat` production in
the Lua grammar:

    stat ::=  varlist `=´ explist |
              ...

Hopefully won't be too hard. Looks like I'm just failing to detect
termination of the statement correctly.