feat: embed kind, line and column info in comment nodes
In the long run probably want to do this for all nodes, but for now we
only _need_ to do it for comment nodes.

This will enable us to build comment-related features in the future,
which we'll need because the whole point of docvim is to turn comments
into docs. ie. with this info, we'll be able to turn:

    --[
    --
    -- @header
    --
    -- # Heading
    --
    -- ## Subheading
    --
    -- Docs. A link |target|.
    --

Into:

    @header

    # Heading

    ## Subheading

    Docs. A link |target|.

And from there into a structured representation.

Ditto for something like:

    --[==[[

    @header

    # Heading

    ## Subheeading

    Docs. A link |target|.

    ]==]

Doing this for all nodes (or at least, all expression and statement
nodes) will be verbose and fiddly, so not making it a priority for now.