feat: add Location info to Node struct
Originally wanted to do this in the variant enum variants of ExpKind and
StatementKind (ie. make them all have a `location` property), but that
doesn't work. So, instead put it in Node.

Updating the tests was a pain, so I converted them to snapshot tests.
Instead, testing the error case, which is better suited to inline tests.